refactor: changed listings to bookings on the user page
