+++ src/components/BookingCard/index.ts
@@ -1,0 +1,1 @@
+export * from './BookingCard';
