--- src/components/ListingCard/index.ts
@@ -1,1 +1,0 @@
-export * from './ListingCard';
